accessible/tests/mochitest/test_actions_doc.html

<html>

<head>
  <title>nsIAccessible actions testing for document</title>

  <link rel="stylesheet" type="text/css"
        href="chrome://mochikit/content/tests/SimpleTest/test.css" />

  <script type="application/javascript"
          src="chrome://mochikit/content/MochiKit/packed.js"></script>
  <script type="application/javascript"
          src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js"></script>

  <script type="application/javascript"
          src="chrome://mochikit/content/a11y/accessible/common.js"></script>
  <script type="application/javascript"
          src="chrome://mochikit/content/a11y/accessible/nsIAccessible_actions.js"></script>

  <script type="application/javascript">
    function doTest()
    {
      var docAcc = getAccessible(document, [nsIAccessibleDocument]);
      
      if (!docAcc) {
        SimpleTest.finish();
        return;
      }

      is(docAcc.numActions, 1, "Wrong number of actions for document!");
      is(docAcc.getActionName(0), "click", "Wrong action name for document!");
      
      SimpleTest.finish();
    }

    SimpleTest.waitForExplicitFinish();
    addLoadEvent(doTest);
  </script>
</head>

<body role="button">

  <a target="_blank"
     href="https://bugzilla.mozilla.org/show_bug.cgi?id=473732"
     title="Related to work on bug 473732">
    Mozilla Bug 473732
  </a>
  <p id="display"></p>
  <div id="content" style="display: none"></div>
  <pre id="test">
  </pre>

</body>
</html>
